crs-dr-octopus-master/dr-octopus/requirements.txt
flask
flask_restplus
requests
Flask-API
pyopenssl
python-crontab
